pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw"
version = "0.1.0"
description = "Local AI agent powered by Ollama, accessible via Microsoft Teams"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "botbuilder-core>=4.14.0",
    "botbuilder-integration-aiohttp>=4.14.0",
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "chromadb>=0.4.22",
    "duckduckgo-search>=4.4.0",
    "unstructured>=0.12.0",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
